fix: pr comment

+++ pkg/prometheus/format.go
@@ -7,11 +7,11 @@ import (
 )
 
 func FormatMilliCPU(cores float64) *resource.Quantity {
-	return resource.NewMilliQuantity(int64(float64(cores)*1000), resource.DecimalSI)
+	return resource.NewMilliQuantity(int64(cores*1000), resource.DecimalSI)
 }
 
 // FormatMiB converts a float64 byte count from a PromQL query result into a resource.Quantity rounded to the nearest MiB.
 // The maximum rounding error is 0.5 MiB.
 func FormatMiB(bytes float64) *resource.Quantity {
-	return resource.NewQuantity(int64(math.Round(float64(bytes)/1024/1024)*1024*1024), resource.BinarySI)
+	return resource.NewQuantity(int64(math.Round(bytes/1024/1024)*1024*1024), resource.BinarySI)
 }
